LOOOOOOL

+++ src/calloc.c
@@ -6,7 +6,7 @@
 /*   By: pbremond <pbremond@student.42nice.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/28 16:33:15 by pbremond          #+#    #+#             */
-/*   Updated: 2024/03/15 15:11:12 by pbremond         ###   ########.fr       */
+/*   Updated: 2024/04/15 14:38:03 by pbremond         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,7 +34,7 @@ void	*CALLOC(size_t nb_elem, size_t size)
 		void *mem = MALLOC(nb_elem * size);
 		if (!mem)	// errno already set in malloc
 			return NULL;
-		if (size > opt.small_alloc_max_sz)
+		if (size > (size_t)opt.small_alloc_max_sz)
 			ft_bzero(mem, nb_elem * size);	// TODO: optimize with 64-bit wide call
 		return mem;
 	}
